modified:   src/ws.c

+++ src/ws.h
@@ -19,7 +19,7 @@
 #define WS_FRAME_PONG 0x8A
 #define WS_FRAME_CLOSE 0x88
 
-int preamble_ws_frame(char **frame_data, size_t length, unsigned char type);
+size_t preamble_ws_frame(char **frame_data, size_t length, unsigned char type);
 #endif
 
 typedef struct
